kubernetes DistributedCache - WIP
