+++ jcachex-core/src/main/java/io/github/dhruv1110/jcachex/profiles/ProfileName.java
@@ -30,8 +30,6 @@ public enum ProfileName {
             ProfileCategory.ADVANCED),
     ZERO_COPY("ZERO_COPY", "Zero-copy optimized cache for minimal memory allocation", ProfileCategory.ADVANCED),
     HARDWARE_OPTIMIZED("HARDWARE_OPTIMIZED", "Hardware-optimized cache leveraging CPU-specific features",
-            ProfileCategory.ADVANCED),
-    KUBERNETES_DISTRIBUTED("KUBERNETES_DISTRIBUTED", "Kubernetes distributed cache optimized for cluster environments",
             ProfileCategory.ADVANCED);
 
     private final String value;
@@ -133,7 +131,7 @@ public enum ProfileName {
      * @return array of advanced profiles
      */
     public static ProfileName[] getAdvancedProfiles() {
-        return new ProfileName[] { ML_OPTIMIZED, ZERO_COPY, HARDWARE_OPTIMIZED, KUBERNETES_DISTRIBUTED };
+        return new ProfileName[] { ML_OPTIMIZED, ZERO_COPY, HARDWARE_OPTIMIZED };
     }
 
     @Override
