+++ jcachex-spring/src/main/resources/META-INF/spring.factories
@@ -1,4 +1,3 @@
 # Auto-Configuration
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  io.github.dhruv1110.jcachex.spring.configuration.JCacheXAutoConfiguration,\
-  io.github.dhruv1110.jcachex.spring.distributed.JCacheXDistributedAutoConfiguration
+  io.github.dhruv1110.jcachex.spring.configuration.JCacheXAutoConfiguration
